Refactor root and schema into their own files
